+++ Samples~/AssemblyInfo.cs
@@ -1,9 +1,0 @@
-﻿/*
-Yarn Spinner is licensed to you under the terms found in the file LICENSE.md.
-*/
-
-using System.Reflection;
-
-[assembly: AssemblyVersion("3.0.0.235")]
-[assembly: AssemblyFileVersion("3.0.0.235")]
-[assembly: AssemblyInformationalVersion("3.0.0-beta2.Branch.release/v3.0.0-beta2.Sha.e95b5ff4c7576725366c01ba842f4b0a5e805c76")]