+++ Samples~/Shared Assets/Scripts/Editor/SimpleCharacterEditor.cs
@@ -1,8 +1,0 @@
-using UnityEditor;
-using Yarn.Unity.Editor;
-
-namespace Yarn.Unity.Samples.Editor
-{
-    [CustomEditor(typeof(SimpleCharacter))]
-    public class SimpleCharacterEditor : YarnEditor { }
-}